+++ NEWS.md
@@ -3,9 +3,13 @@ title: "NEWS"
 output: html_document
 ---
 
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
+
+## backbone 1.1.1
+
+* remove `poibin` package import
+* update `sdsm()` to function without using `poibin` package
+* use source code from `poibin` under GPL-2, Yili Hong (2013)
+
 ## backbone 1.1.0
 
 * add support for sparse matrices